+++ server/model/todo.go
@@ -4,6 +4,6 @@ type Todo struct {
 	ID        string `json:"id"`
 	Title     string `json:"title"`
 	Completed bool   `json:"completed"`
-	Order     int    `json:"order"`
+	Order     int64  `json:"order"`
 	URL       string `json:"url"`
 }
